requirements.txt
geopandas>=0.14
shapely>=2.0
pandas>=2.0
pyarrow>=14.0
PyYAML>=6.0
